server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
openpyxl>=3.1.5
PyJWT>=2.8.0
httpx>=0.27.0
cryptography>=42.0.0
